+++ work_spacer/lib/models/room.dart
@@ -1,19 +1,26 @@
 import 'package:flutter/foundation.dart';
+import 'package:json_annotation/json_annotation.dart';
 import 'package:work_spacer/models/workspace.dart';
 
+part 'room.g.dart';
+
+@JsonSerializable()
 @immutable
 class Room extends Workspace {
   final int capacity;
-  final bool hasProjector;
-  final bool hasWhiteboard;
-  final bool hasTeleconference;
+  final bool projector;
+  final bool whiteboard;
+  final bool teleconferenceSystem;
 
   const Room({
     required super.id,
     required super.floor,
     required this.capacity,
-    required this.hasProjector,
-    required this.hasWhiteboard,
-    required this.hasTeleconference,
+    required this.projector,
+    required this.whiteboard,
+    required this.teleconferenceSystem,
   });
+
+  factory Room.fromJson(Map<String, dynamic> json) => _$RoomFromJson(json);
+  Map<String, dynamic> toJson() => _$RoomToJson(this);
 }