+++ work_spacer/lib/stores/rooms_store.dart
@@ -6,6 +6,8 @@ import 'package:work_spacer/models/workspace.dart';
 import 'package:work_spacer/stores/block_store.dart';
 import 'package:work_spacer/stores/filter_store.dart';
 
+import '../src/helpers/proxy.dart';
+
 part 'rooms_store.g.dart';
 
 class RoomsStore = _RoomsStore with _$RoomsStore;
@@ -36,11 +38,11 @@ abstract class _RoomsStore with Store {
           final isCapacityValid =
               _checkFilterLessOrEqual(FilterParameter.capacity, room.capacity);
           final isWhiteboardValid =
-              _checkFilterEqual(FilterParameter.whiteboard, room.hasWhiteboard);
+              _checkFilterEqual(FilterParameter.whiteboard, room.whiteboard);
           final isProjectorValid =
-              _checkFilterEqual(FilterParameter.projector, room.hasProjector);
+              _checkFilterEqual(FilterParameter.projector, room.projector);
           final isTeleconferenceValid = _checkFilterEqual(
-              FilterParameter.teleconference, room.hasTeleconference);
+              FilterParameter.teleconference, room.teleconferenceSystem);
 
           final isAvailable = _checkAvailability();
 
@@ -123,18 +125,33 @@ abstract class _RoomsStore with Store {
     //TODO handle backend
     //IMPORTANT! we need to limit this list based on the role (see User.dart -> enum), so strapi model should implement something like a list of roles that can acces each workspace,
     // e.g. Room #3 should have a list [Mid, Senior] and when fetching data here, only fetch workspaces that the user can actually access
-    await Future.delayed(const Duration(milliseconds: 500));
-    _rooms = roomsDummy..sort((room1, room2) => room1.id.compareTo(room2.id));
+    // await Future.delayed(const Duration(milliseconds: 500));
+
+    var res = await Proxy.data('conference-rooms');
+    var jsonRooms = res['data'].map((e) {
+      return <String, dynamic>{"id": e['id'], ...e['attributes']};
+    }).toList();
+
+    var rooms =
+        jsonRooms.map<Room>((e) => Room.fromJson(e)).toList() as List<Room>;
+
+    _rooms = rooms..sort((room1, room2) => room1.id.compareTo(room2.id));
     inProgress = false;
   }
 
   @action
   reserveRoom(int? userId, Workspace roomAsWorkspace, DateTime date,
-      TimeOfDay time, int hours) {
+      TimeOfDay time, int hours) async {
     if (userId == null) {
       return;
     }
     final Room room = roomAsWorkspace as Room;
-    //TODO handle backend
+    var combinedDate = DateTime(date.year, date.month, date.day, time.hour, time.minute);
+    var res = await Proxy.data('cr-reservation', method: "POST", body: {
+      "startDate": combinedDate.toIso8601String(),
+      "duration": hours,
+      "idDesk": room.id,
+      "idEmployee": userId
+    });
   }
 }