+++ CHANGELOG.md
@@ -5,9 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [0.14.0] - 2022-06-19
+### Added
+* Proxy connection to firebase
+* Data fetch for items
+* JsonSerialisation and Deserialisation 
+* Log in function
+### Changed
+* Local data storages functions
 
 # [0.13.0] - 2022-06-14
-# Added
+### Added
 * Firebase functions
 * Firebase create user function
 * Strapi employee and admin lifecycle webhook